Bug 149982 - Fix diagnosis of UI thread interrupt
The original fix for bug 149982 was intended to
address UI thread interrupts interfering with the
timed wait step in the Lock::uiSafeAcquire method.
It was intended then that interrupts of the job-rule
waiting phase should be recognized, which thanks to
changes in Eclipse Platform they now are (they were
not before).  This requires changes to allow
diagnosis of this specific situation and an update
to the failing JUnit test to assert that this is
what causes the transaction lock acquisition to fail.

Change-Id: Icd95e71f513c1458f8385d99e886924526bb7c9a
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>